@@ -2,49 +2,67 @@
   "name": "@hoprnet/hopr-ethereum",
   "version": "0.0.1",
   "description": "On-chain logic for hopr.network",
-  "main": "index.ts",
   "homepage": "https://hopr.network",
   "repository": {
     "type": "git",
     "url": "https://github.com/hoprnet/hopr-ethereum.git"
-  },  
+  },
   "license": "LGPL-3.0-only",
+  "private": false,
+  "main": "./build/lib/scripts/api.js",
+  "files": [
+    "truffle-networks.json",
+    "truffle-config.js",
+    "contracts/*.sol",
+    "build/extracted/",
+    "build/lib/"
+  ],
   "engines": {
-    "node": "~12"
+    "yarn": "^1.19.2",
+    "node": "^12"
+  },
+  "scripts": {
+    "build": "ts-node -T ./scripts/cli.ts build",
+    "coverage": "ts-node -T ./scripts/cli.ts coverage",
+    "fund": "ts-node -T ./scripts/cli.ts fund",
+    "migrate": "ts-node -T ./scripts/cli.ts migrate",
+    "network": "ts-node -T ./scripts/cli.ts network",
+    "test": "ts-node -T ./scripts/cli.ts test",
+    "format": "prettier --check ."
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "yarn format && yarn build && git add ."
+    }
   },
   "dependencies": {
-    "@openzeppelin/contracts": "^2.4.0"
+    "@hoprnet/hopr-demo-seeds": "https://github.com/hoprnet/hopr-demo-seeds.git",
+    "@hoprnet/hopr-utils": "^0.0.2-dev.119b5cf0b",
+    "@openzeppelin/contracts": "^3.0.1",
+    "@openzeppelin/test-helpers": "^0.5.5",
+    "@truffle/debug-utils": "^4.1.1",
+    "@truffle/hdwallet-provider": "^1.0.34",
+    "bignumber.js": "^9.0.0",
+    "ganache-core": "^2.10.2",
+    "truffle": "^5.1.23",
+    "truffle-plugin-verify": "^0.3.10",
+    "ts-node": "^8.9.1",
+    "web3": "^1.2.7"
   },
   "devDependencies": {
-    "@openzeppelin/test-helpers": "^0.5.4",
-    "@types/chai": "^4.2.7",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^12.12.17",
+    "@types/chai": "^4.2.11",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^13.13.4",
     "chai": "^4.2.0",
-    "ganache-cli": "^6.7.0",
-    "mocha": "^6.2.2",
-    "solhint": "^2.3.0",
-    "solidity-coverage": "^0.6.7",
-    "ganache-cli-coverage": "https://github.com/frangio/ganache-cli/releases/download/v6.4.1-coverage/ganache-cli-coverage-6.4.1.tgz",
-    "truffle": "^5.1.3",
+    "husky": "^4.2.5",
+    "mocha": "^7.1.2",
+    "prettier": "^2.0.5",
+    "prettier-plugin-solidity": "^1.0.0-alpha.49",
+    "solhint": "^2.3.1",
+    "solidity-coverage": "^0.7.4",
     "truffle-typings": "^1.0.8",
-    "ts-node": "^8.5.4",
-    "typechain": "^1.0.3",
+    "typechain": "^1.0.5",
     "typechain-target-truffle": "^1.0.2",
-    "typescript": "^3.7.3"
-  },
-  "files": [
-    "lib/"
-  ],
-  "scripts": {
-    "postinstall": "yarn build",
-    "lint:sol": "solhint 'contracts/**/*.sol'",
-    "build:tsc": "tsc",
-    "build:types": "typechain --target truffle --outDir './types/truffle-contracts' './build/**/*.json' && scripts/fix.sh",
-    "build:sol": "yarn lint:sol && truffle compile && yarn build:types",
-    "build": "yarn build:sol && yarn build:tsc",
-    "migrate": "yarn build && truffle migrate",
-    "test": "yarn build && scripts/test.sh",
-    "coverage": "yarn build && scripts/coverage.sh"
+    "typescript": "^3.8.3"
   }
 }
